refactor(menu): move SubmenuToggle to deprecated folder
Move SubmenuToggle component to deprecated folder and update imports.
The deprecated SubmenuToggle uses aliased destructuring to maintain
backward compatibility with the renamed context variables:
- submenuTriggerId: submenuToggleId
- submenuTriggerRef: submenuToggleRef
- isHoveringSubmenuTriggerRef: isHoveringSubmenuToggleRef
