+++ packages/react/src/menu/deprecated/styles.js
@@ -1,0 +1,10 @@
+const useSubmenuToggleStyle = () => {
+  return {
+    cursor: 'pointer',
+    display: 'inline-flex',
+  };
+};
+
+export {
+  useSubmenuToggleStyle,
+};
